add: update documentation to the project. docker-compose support
